Add findOpcode() function to envi/IMemory, which gets exposed to getTrace().
This allows you to specify an address (or uses EIP if not specified) and an opcode mnemonic, and will loop over the assembly until it comes to your opcode, then returns it.

You can also specify a verbose flag, and it will print out all the disassembly between you and your opcode.